+++ app/tracklist/src/main/java/com/arsnyan/tracklist/network/repository/TrackDataSource.kt
@@ -5,5 +5,5 @@ import com.arsnyan.tracklist.network.model.Track
 interface TrackDataSource {
     suspend fun getAllTracks(): Result<List<Track>>
     suspend fun searchTracks(query: String): Result<List<Track>>
-    suspend fun getTrackById(id: Int): Result<Track>
+    suspend fun getTrackById(id: Long): Result<Track>
 }
